simplify tests by isolating the `exec` call
This was “solved” far too cleverly before: invocations of `run` which
would have successfully proceeded to the `exec` were prevented from
doing so by throwing an exception from the overridden `create` and
`open` methods. This also meant that the success case in the overridden
method had to be to throw an exception.

By isolating the `exec` call in another method, it can be overridden
separately, which makes the control flow much more obvious. It will also
allow testing the `exec` call itself, which currently has a bug.