fix application of PerlIO layers in constructor
We cannot use `open()` here because its `+<` mode is just O_RDWR and
its `+>` mode is O_RDWR|O_CREAT|O_TRUNC, neither of which works for us:
what we want is O_RDWR|O_CREAT, which is only expressible by `sysopen`.

But contrary to my mental model when I wrote the code, using `binmode()`
to modify layers after the fact has a very different effect from passing
the same layer specification to `open()`.

And the only way to get that behaviour is to pass that specification
while opening the handle – for which `sysopen` has no actual argument.
It can be done – but only by `use open`.

Which means `eval()`. Awful, but not past the hold-your-nose threshold.